# server/v2-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Emotion Machine API",
    "description": "API for building AI companions with persistent relationships",
    "version": "2.1.0"
  },
  "servers": [
    {
      "url": "https://api.emotionmachine.ai",
      "description": "Production"
    },
    {
      "url": "http://localhost:8100",
      "description": "Local development"
    }
  ],
  "paths": {
    "/v1/companions": {
      "get": {
        "tags": [
          "client-api"
        ],
        "summary": "List Companions",
        "operationId": "list_companions_v1_companions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompanionListItem"
                  },
                  "type": "array",
                  "title": "Response List Companions V1 Companions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "ProjectApiKey": []
          }
        ]
      },
      "post": {
        "tags": [
          "client-api"
        ],
        "summary": "Create Companion",
        "operationId": "create_companion_v1_companions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionResource"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ProjectApiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanionConfig-Input": {
        "properties": {
          "system_prompt": {
            "$ref": "#/components/schemas/SystemPrompt"
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryConfig"
          },
          "knowledge": {
            "$ref": "#/components/schemas/KnowledgeConfig"
          },
          "inference": {
            "$ref": "#/components/schemas/InferenceConfig"
          },
          "voice": {
            "$ref": "#/components/schemas/app__models__companion__VoiceConfig"
          },
          "classifier": {
            "$ref": "#/components/schemas/ClassifierConfig"
          },
          "context_mode": {
            "type": "string",
            "enum": [
              "legacy",
              "layered",
              "raw"
            ],
            "title": "Context Mode",
            "default": "layered"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerAttachment"
            },
            "type": "array",
            "title": "Layers"
          },
          "context": {
            "$ref": "#/components/schemas/ContextSettings"
          }
        },
        "type": "object",
        "title": "CompanionConfig"
      },
      "CompanionConfig-Output": {
        "properties": {
          "system_prompt": {
            "$ref": "#/components/schemas/SystemPrompt"
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryConfig"
          },
          "knowledge": {
            "$ref": "#/components/schemas/KnowledgeConfig"
          },
          "inference": {
            "$ref": "#/components/schemas/InferenceConfig"
          },
          "voice": {
            "$ref": "#/components/schemas/VoiceConfig-Output"
          },
          "classifier": {
            "$ref": "#/components/schemas/ClassifierConfig"
          },
          "context_mode": {
            "type": "string",
            "enum": [
              "legacy",
              "layered",
              "raw"
            ],
            "title": "Context Mode",
            "default": "layered"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerAttachment"
            },
            "type": "array",
            "title": "Layers"
          },
          "context": {
            "$ref": "#/components/schemas/ContextSettings"
          }
        },
        "type": "object",
        "title": "CompanionConfig"
      },
      "CompanionCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanionConfig-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CompanionCreateRequest"
      },
      "CompanionListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "last_updated": {
            "type": "string",
            "title": "Last Updated"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "last_updated",
          "project_id"
        ],
        "title": "CompanionListItem"
      },
      "CompanionResource": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "config": {
            "$ref": "#/components/schemas/CompanionConfig-Output"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "config",
          "created_at",
          "project_id"
        ],
        "title": "CompanionResource"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InferenceConfig": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "LLM model ID (e.g., 'claude-sonnet-4.5', 'openai-gpt4o', 'gemini-2.5-flash')"
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          },
          "max_output_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 16384.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Output Tokens"
          }
        },
        "type": "object",
        "title": "InferenceConfig",
        "description": "Unified LLM settings - used for both text API and voice sessions."
      },
      "KnowledgeConfig": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Whether knowledge retrieval is enabled",
            "default": false
          },
          "gate_strategy": {
            "type": "string",
            "title": "Gate Strategy",
            "description": "Gate strategy: 'keyword', 'always', or 'llm'",
            "default": "keyword"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "description": "Max knowledge items to retrieve",
            "default": 5
          },
          "min_confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Confidence",
            "description": "Min confidence score",
            "default": 0.0
          },
          "classifier_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classifier Summary",
            "description": "Custom description for the classifier to understand when to enable this layer. E.g., 'Health information about menstrual cycles, symptoms, and period tracking'"
          }
        },
        "type": "object",
        "title": "KnowledgeConfig",
        "description": "Knowledge/RAG configuration for the companion."
      },
...
      "MemoryConfig": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "version": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 1.0,
            "title": "Version",
            "description": "Memory version: 1=vector retrieval, 2=scratchpad",
            "default": 1
          },
          "core_memories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Core Memories"
          },
          "memory_evaluation_prompt": {
            "type": "string",
            "title": "Memory Evaluation Prompt",
            "default": ""
          },
          "recency": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Recency",
            "default": 0.995
          },
          "top_k": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 50
          },
          "min_saliency": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Saliency",
            "default": 0.2
          },
          "max_entries": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Max Entries",
            "description": "Max scratchpad entries for v2",
            "default": 100
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "LLM model for v2 ingestion"
          },
          "ingestion_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingestion Prompt",
            "description": "Custom ingestion prompt for v2"
          }
        },
        "type": "object",
        "title": "MemoryConfig"
      },
...
      "SystemPrompt": {
        "properties": {
          "full_system_prompt": {
            "type": "string",
            "title": "Full System Prompt",
            "description": "Complete system prompt",
            "default": ""
          },
          "identity": {
            "type": "string",
            "title": "Identity",
            "description": "Character identity",
            "default": ""
          },
          "personality": {
            "type": "string",
            "title": "Personality",
            "description": "Character personality",
            "default": ""
          },
          "style": {
            "type": "string",
            "title": "Style",
            "description": "Communication style",
            "default": ""
          },
          "backstory": {
            "type": "string",
            "title": "Backstory",
            "description": "Character backstory",
            "default": ""
          },
          "additional_instructions": {
            "type": "string",
            "title": "Additional Instructions",
            "description": "Additional instructions",
            "default": ""
          },
          "self_image": {
            "type": "string",
            "title": "Self Image",
            "description": "Character self-image",
            "default": ""
          }
        },
        "type": "object",
        "title": "SystemPrompt"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Clerk JWT token"
      },
      "ProjectApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "APIKey",
        "description": "Use `Authorization: Bearer <PROJECT_API_KEY>`"
      }
    }
  }
}